refactor(recipe): dropzone component.
create own dropzone component for add and edit recipe.

refactor recipe-features.
